Pawn structure eval
Pawn are evaluated and added to the evaluation at the end. To speed this up a pawn hash table is used (added pawn structure zobrist key to position class).
The hash table has around a 95-98% hitrate (with a 4 MB size), as pawn structure does not change that much during search.
Gained +80 elo compared to previous commit.