Mate finding tests and bug fixes
Testing revealed that previously a partial result could override a decisive mate score. In reality partial results should not be used to override decisive results.
That is now fixed, so only non-mate scores can be overridden by a better move found in a partial search. Difference in elo rating is negligible.